refactor: remove block translation utilities and shared block types
- Deleted the blockTranslation.ts file containing translation utilities for journey blocks.
- Removed the block.ts file which defined shared interfaces and types for various block types.
- Updated related schema files to ensure proper integration of block types without the removed utilities.
