+++ apis/api-journeys-modern/src/schema/block/radioQuestion/radioQuestion.ts
@@ -1,7 +1,7 @@
 import { builder } from '../../builder'
 import { Block } from '../block'
 
-builder.prismaObject('Block', {
+export const RadioQuestionBlock = builder.prismaObject('Block', {
   interfaces: [Block],
   variant: 'RadioQuestionBlock',
   isTypeOf: (obj: any) => obj.typename === 'RadioQuestionBlock',
