+++ apis/api-journeys-modern/src/schema/journeyAiTranslate/getCardBlocksContent/getRadioQuestionBlockContent.spec.ts
@@ -5,13 +5,11 @@ import { getRadioQuestionBlockContent } from './getRadioQuestionBlockContent'
 describe('getRadioQuestionBlockContent', () => {
   const radio1: Partial<Block> = {
     id: 'r1',
-    typename: 'RadioQuestionBlock',
-    label: 'Q1'
+    typename: 'RadioQuestionBlock'
   }
   const radio2: Partial<Block> = {
     id: 'r2',
-    typename: 'RadioQuestionBlock',
-    label: 'Q2'
+    typename: 'RadioQuestionBlock'
   }
   const notRadio: Partial<Block> = {
     id: 'n1',
@@ -19,30 +17,31 @@ describe('getRadioQuestionBlockContent', () => {
     label: 'NotQ'
   }
 
-  it('includes block id and label in output', () => {
+  it('includes block id in output', () => {
     const result = getRadioQuestionBlockContent({
       blocks: [radio1 as Block],
       block: radio1 as Block
     })
     expect(result).toContain('Block ID: r1')
-    expect(result).toContain('Text: Q1')
   })
 
-  it('lists all radio question block labels from blocks', () => {
+  it('includes radio question blocks without labels', () => {
     const result = getRadioQuestionBlockContent({
       blocks: [radio1 as Block, radio2 as Block],
       block: radio1 as Block
     })
-    expect(result).toContain('- Q1')
-    expect(result).toContain('- Q2')
+    expect(result).toContain('### Questions:')
+    // Should not contain any label text since labels are removed
+    expect(result).not.toContain('Q1')
+    expect(result).not.toContain('Q2')
   })
 
-  it('does not list non-radio-question blocks', () => {
+  it('does not include non-radio-question blocks', () => {
     const result = getRadioQuestionBlockContent({
       blocks: [radio1 as Block, notRadio as Block],
       block: radio1 as Block
     })
-    expect(result).toContain('- Q1')
+    expect(result).toContain('### Questions:')
     expect(result).not.toContain('NotQ')
   })
 
@@ -51,7 +50,8 @@ describe('getRadioQuestionBlockContent', () => {
       blocks: [radio1 as Block],
       block: radio1 as Block
     })
-    expect(result).toContain('- Q1')
+    expect(result).toContain('Block ID: r1')
+    expect(result).toContain('### Questions:')
   })
 
   it('handles empty blocks array', () => {
@@ -60,6 +60,6 @@ describe('getRadioQuestionBlockContent', () => {
       block: radio1 as Block
     })
     expect(result).toContain('Block ID: r1')
-    expect(result).toContain('Text: Q1')
+    expect(result).toContain('### Questions:')
   })
 })
